Back porting coverity issues fix patch from master branch. Fix coverity issue:1111387 Explicit null dereferenced Fix coverity issue:1111388 Explicit null dereferenced Fix coverity issue:1111411 Missing break in switch Fix coverity issue:1111412 Missing break in switch Fix coverity issue:1111450 Resource leak Fix coverity issue:1111451 Resource leak Fix coverity issue:1111472 Dereference before null check Fix coverity issue:1111473 Dereference before null check Fix coverity issue:1111474 Dereference before null check Fix coverity issue:1111475 Wrong sizeof argument Fix coverity issue:1111495 Use after free Fix coverity issue:1111496 Use after free Fix coverity issue:1222989 Bad bit shift operation Fix coverity issue:1222990 Bad bit shift operation Fix coverity issue:1222994 Explicit null dereferenced

+++ pool_process_query.c
@@ -519,7 +519,7 @@ POOL_STATUS pool_process_query(POOL_CONNECTION *frontend,
 /* used only in pool_parallel_exec */
 #define BITS (8 * sizeof(long int))
 
-static void set_fd(unsigned long fd ,unsigned long *setp)
+static void set_fd(int fd ,unsigned long *setp)
 {
 	unsigned long tmp = fd / FD_SETSIZE;
 	unsigned long rem = fd % FD_SETSIZE;
@@ -527,7 +527,7 @@ static void set_fd(unsigned long fd ,unsigned long *setp)
 }
 
 /* used only in pool_parallel_exec */
-static int isset_fd(unsigned long fd, unsigned long *setp)
+static int isset_fd(int fd, unsigned long *setp)
 {
 	unsigned long tmp = fd / FD_SETSIZE;
 	unsigned long rem = fd % FD_SETSIZE;
